Default content-type, status summary, clean-up
The default content-type is now "application/json", if no "Accept" header
is provided.

An additional field "connectionStateSummary" is added. It contains
"enabled", "disabled", or "mixed". The logic, which of those states
constitutes an error or not, is left to the calling application.

REST services for adding, deleting, and updating connections are
removed, since they do not make sense for a monitoring scenario.